Made inventory

+++ Assets/Scripts/Player/PlayerController.cs
@@ -11,16 +11,14 @@ using UnityEngine.InputSystem;
 public class PlayerController : MonoBehaviour
 {
     public bool exposed = false; 
-    public int money = 0;
-    public int maxMoney = 100;
+    public float money = 0f;
+    public float maxMoney = 100f;
     public float speed = 4f;
     public float sprintMultiplier = 1.5f;
     public float crouchMultiplier = 0.8f;
     public bool movementRestriction = false;
     public int Stamina = 100;
     private int StaminaMod;
-    
-    
 
     private Vector2 moveDir; // used for WASD movement
     public Vector2 lastDir;
